requirements.txt
attrs==22.2.0
aws-cdk-lib==2.64.0
aws-cdk.asset-awscli-v1==2.2.64
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.53
boto3==1.26.69
botocore==1.29.69
cattrs==22.2.0
constructs==10.1.247
exceptiongroup==1.1.0
jmespath==1.0.1
jsii==1.74.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
